Added standard type/vararg headers, kprintf now properly supports varargs

+++ src/multiboot.h
@@ -1,6 +1,8 @@
 #ifndef __MULTIBOOT_HEADER__
 #define __MULTIBOOT_HEADER__
 
+#include "stdtypes.h"
+
 // Multiboot header info
 #define     MULTIBOOT_HEADER_MAGIC      0x1BADB002
 #define     MULTIBOOT_HEADER_FLAGS      0x00010003
@@ -14,16 +16,16 @@
 
 typedef struct multiboot_info
 {
-    unsigned long flags;
-    unsigned long mem_lower;
-    unsigned long mem_upper;
-    unsigned long boot_device;
-    unsigned long cmdline;
-    unsigned long mods_count;
-    unsigned long mods_addr;
+    uint32_t flags;
+    uint32_t mem_lower;
+    uint32_t mem_upper;
+    uint32_t boot_device;
+    uint32_t cmdline;
+    uint32_t mods_count;
+    uint32_t mods_addr;
     unsigned char syms[16];
-    unsigned long nmap_length;
-    unsigned long nmap_addr;
+    uint32_t nmap_length;
+    uint32_t nmap_addr;
     
 } multiboot_info_t;
